forseti-meshd/requirements.txt
fastapi>=0.135.0
uvicorn>=0.44.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.18.0
httpx>=0.28.0
pynacl>=1.6.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
